school_management_system/requirements.txt
# Django and core packages
Django>=5.0,<6.1
djangorestframework>=3.14.0
django-ninja>=1.0.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
argon2-cffi>=23.1.0  # Secure password hashing

# CORS
django-cors-headers>=4.3.0

# Filtering
django-filter>=23.5

# Database (PostgreSQL support - optional)
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0

# Development
ipython>=8.0.0

# Production (optional)
gunicorn>=21.2.0
whitenoise>=6.6.0
